Slight enhancements to ball, bumpers, names
I'm still working out how to get the bumpers to add score. I think the easiest way to do it is by just having separate scripts for each bumper type (10,50,100) that add score but I don't understand how your score system works. I was trying to call GameManager and have it trigger on collision but for some reason it won't collide. Perhaps something to do with the 2D?